[#15] improving first test with dpckan_test module
